+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Improve native executable performance by utilizing Profile-Guided Optimization.
+
 # 2.1.5
 
 - Fix Windows application not being able to get output from parser when `DEBUG` environment variable is set ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/1513)).